services/ai-ml/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
transformers==4.36.0
torch==2.1.1
google-generativeai==0.3.2
openai==1.3.7
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-json-logger==2.0.7
prometheus-client==0.19.0
mlflow==2.8.1
joblib==1.3.2
nltk==3.8.1
spacy==3.7.2
